Expose WAL cache and replication stream pos in PostgresConnector API
The PostgresConnector serves as an entry point to starting the
supervision tree of processes that are connected to Postgres or
carry out related tasks.

This change is an attempt to start shaping up a "facade API" in the
PostgresConnector module to abstract away internal details of the
modules it uses in its implementation.