Fix an issue when decoding from Json
When using :atoms! decoding fails occasionally due to the
:included_tables atom not being defined. I'm guessing this
happens because of dynamic module loading in beam. If that's
true, then it won't be an issue in a release. But in order
to be able to run the app in development, we relax the decoder
to allow it to call String.to_atom() in this case.